+++ cypress/support/step_definitions/html.js
@@ -30,13 +30,13 @@ Then('I set on {string} text {string}', (templateSelector, templateValue) => {
 });
 
 Then('I expect {string} is closed', (selector) => {
-  cy.get(selector, { timeout: 10000 }).should('not.exist');
+  cy.get(selector, { timeout: 60000 }).should('not.exist');
 });
 
 Then('I expect {string} toast to appear with text {string}', (type, templateExpectedValue) => {
   const expectedValue = nunjucks.renderString(templateExpectedValue, cy.context);
 
-  cy.get('.q-notification', { timeout: 10000 })
+  cy.get('.q-notification', { timeout: 60000 })
     .should('exist')
     .and('have.class', `bg-${type}`)
     .within(() => {